Move analytics plan policy helper out of plugins
