+++ templates/analytics/server/lib/agent-chat-plan-mode.ts
